requirements.txt
flask==2.3.3
PyPDF2==3.0.1
langchain==0.3.27
langchain-openai==0.3.28
langchain-community==0.3.30
faiss-cpu==1.12.0
openai==1.109.1
requests==2.32.5
pydantic==2.11.7
pydantic-core==2.33.2
SQLAlchemy==2.0.41
langchain-core==0.3.78
langchain-text-splitters==0.3.9
langsmith==0.4.8
PyYAML==6.0.2
numpy==2.2.1
python-dotenv==1.0.1
pinecone==5.3.1
langchain-pinecone==0.1.3
